pyproject.toml
[project]
name = "order-book-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "Aidar Rakhmatullin",email = "nedopolz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "binance-futures-connector (>=4.1.0,<5.0.0)",
    "sortedcontainers (>=2.4.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
py_version = 312
line_length = 119
skip_gitignore = true
group_by_package = true
include_trailing_comma = true
force_sort_within_sections = true
combine_as_imports = true
case_sensitive = true
multi_line_output = 3
indent = 4
atomic = true